+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.3.2 (unreleased)
+0.4.0 (2019-09-11)
 ------------------
 
 - Improve request messages to handle caching